+++ ChangeLog.md
@@ -8,7 +8,6 @@ Features:
  * [#23](https://github.com/iExecBlockchainComputing/PoCo/issues/23): rename createMarketOrder => createOrder || createWorkOrder         [GO/NOGO]
 
 Bugfixes:
- * [#15](https://github.com/iExecBlockchainComputing/PoCo/issues/15): Malicious IexecAPI contract without callback                       [PLANNED in v1.0.13]
 
 
 ### v1.0.13 (unreleased)
@@ -19,8 +18,7 @@ Features:
 
 Bugfixes:
 * [#24](https://github.com/iExecBlockchainComputing/PoCo/issues/24) possibly useless RLC address getter
-* [#20](https://github.com/iExecBlockchainComputing/PoCo/issues/20): add range on trust param ( 0- 100 ) and check it in contracts
-
+* [#15](https://github.com/iExecBlockchainComputing/PoCo/issues/15): Malicious IexecAPI contract without callback 
 
 ### [v1.0.12](https://github.com/iExecBlockchainComputing/PoCo/releases/tag/v1.0.12)
 Initial version. external audit launched on this release v1.0.12